fix: centralise message_sent hook in reply dispatcher
Move message_sent hook from channel-specific delivery paths to the
central createReplyDispatcher(). This fires for ALL channels
(Discord, WhatsApp, Signal, Slack, etc.) without patching each one.

Added optional hookContext to ReplyDispatcherOptions for callers to
provide channel/routing context. Falls back to empty strings when
not provided — plugin still gets the message content.

Removed Discord-specific hook from reply-delivery.ts (now redundant).
Kept deliver.ts hook for non-dispatcher delivery paths (cron, heartbeat).

Refs #6535
